[librpbase] IFst::DirEnt: Change `idx` from `int` to `intptr_t`.
It isn't strictly needed right now, since ISO-9660 is limited to 4 GB,
but it may be needed later for e.g. UDF, and this helps to eliminate
some compiler warnings.

Also re-ordered some variables to reduce padding. (The struct size
remains at 48 bytes on amd64; without re-ordering, it would be 56.)

GcnFst and WiiUFst still use `int` internally, so added casting where
necessary.

IsoPartition: Removed unnecessary `int` casting; cast `size_t` to
`intptr_t` instead of `int`.

+++ src/libromdata/disc/GcnFst.cpp
@@ -460,14 +460,14 @@ const IFst::DirEnt *GcnFst::readdir(IFst::Dir *dirp)
 
 	// Get the directory's fst_entry.
 	RP_D(GcnFst);
-	const GCN_FST_Entry *dir_fst_entry = d->entry(dirp->dir_idx);
+	const GCN_FST_Entry *dir_fst_entry = d->entry(static_cast<int>(dirp->dir_idx));
 	if (!dir_fst_entry) {
 		// dir_idx is corrupted.
 		return nullptr;
 	}
 
 	// Seek to the next entry in the directory.
-	int idx = dirp->entry.idx;
+	int idx = static_cast<int>(dirp->entry.idx);
 	const GCN_FST_Entry *fst_entry = d->entry(idx, nullptr);
 	if (!fst_entry) {
 		// No more entries.
